+++ src/main/java/com/uwetrottmann/trakt5/services/Episodes.java
@@ -23,6 +23,7 @@ import com.uwetrottmann.trakt5.entities.Stats;
 import com.uwetrottmann.trakt5.enums.Extended;
 import retrofit2.Call;
 import retrofit2.http.GET;
+import retrofit2.http.Header;
 import retrofit2.http.Path;
 import retrofit2.http.Query;
 
@@ -60,6 +61,20 @@ public interface Episodes {
             @Query("page") Integer page,
             @Query("limit") Integer limit,
             @Query(value = "extended", encoded = true) Extended extended
+    );
+
+    /**
+     * Variant that allows changing the Cache-Control header.
+     */
+    @GET("shows/{id}/seasons/{season}/episodes/{episode}/comments")
+    Call<List<Comment>> comments(
+            @Path("id") String showId,
+            @Path("season") int season,
+            @Path("episode") int episode,
+            @Query("page") Integer page,
+            @Query("limit") Integer limit,
+            @Query(value = "extended", encoded = true) Extended extended,
+            @Header("Cache-Control") String cacheControl
     );
 
     /**