+++ src/main/java/com/uwetrottmann/trakt5/services/Shows.java
@@ -28,6 +28,7 @@ import com.uwetrottmann.trakt5.enums.Extended;
 import com.uwetrottmann.trakt5.enums.ProgressLastActivity;
 import retrofit2.Call;
 import retrofit2.http.GET;
+import retrofit2.http.Header;
 import retrofit2.http.Path;
 import retrofit2.http.Query;
 
@@ -107,6 +108,18 @@ public interface Shows {
             @Query("page") Integer page,
             @Query("limit") Integer limit,
             @Query(value = "extended", encoded = true) Extended extended
+    );
+
+    /**
+     * Variant that allows changing the Cache-Control header.
+     */
+    @GET("shows/{id}/comments")
+    Call<List<Comment>> comments(
+            @Path("id") String showId,
+            @Query("page") Integer page,
+            @Query("limit") Integer limit,
+            @Query(value = "extended", encoded = true) Extended extended,
+            @Header("Cache-Control") String cacheControl
     );
 
     /**